+++ packages/runtime/plugin-router-v7/src/cli/index.ts
@@ -1,5 +1,6 @@
+import path, { sep } from 'path';
 import type { AppTools, CliPluginFuture } from '@modern-js/app-tools';
-import { logger } from '@modern-js/utils';
+import { isDepExists, logger } from '@modern-js/utils';
 
 export const routerPlugin = (): CliPluginFuture<AppTools> => ({
   name: '@modern-js/plugin-router-v7',
@@ -7,6 +8,7 @@ export const routerPlugin = (): CliPluginFuture<AppTools> => ({
   setup: api => {
     api.config(() => {
       const config = api.getNormalizedConfig();
+      const appContext = api.getAppContext();
       const alias = config?.source?.alias;
       if (typeof alias !== 'undefined') {
         Object.keys(alias).forEach(key => {
@@ -17,19 +19,36 @@ export const routerPlugin = (): CliPluginFuture<AppTools> => ({
           }
         });
       }
+
+      const hasReactRouterDep = isDepExists(
+        appContext.appDirectory,
+        'react-router',
+      );
+
+      const cjs = `${sep}cjs${sep}`;
+      const esm = `${sep}esm${sep}`;
+      const runtimeAlias = hasReactRouterDep
+        ? path.join(appContext.appDirectory, 'node_modules', 'react-router')
+        : require.resolve('../runtime').replace(cjs, esm);
+
+      const finalAlias = {
+        'react-router-dom$': runtimeAlias,
+        '@remix-run/router': runtimeAlias,
+        'react-router-dom/server$': runtimeAlias,
+        [`@${appContext.metaName}/runtime/router/rsc`]: require
+          .resolve('../runtime/rsc')
+          .replace(cjs, esm),
+      };
+
+      if (hasReactRouterDep) {
+        finalAlias['react-router/rsc$'] = require.resolve('react-router/rsc', {
+          paths: [appContext.appDirectory],
+        });
+      }
+
       return {
         resolve: {
-          alias: {
-            'react-router-dom$': require
-              .resolve('../runtime')
-              .replace(/\/cjs\//, '/esm/'),
-            '@remix-run/router': require
-              .resolve('../runtime')
-              .replace(/\/cjs\//, '/esm/'),
-            'react-router-dom/server': require
-              .resolve('../runtime')
-              .replace(/\/cjs\//, '/esm/'),
-          },
+          alias: finalAlias,
         },
         source: {
           globalVars: {