+++ packages/runtime/plugin-runtime/src/cli/template.server.ts
@@ -26,6 +26,7 @@ import {
   createRequestHandler,
 } from '@#metaName/runtime/ssr/server';
 import { RSCServerSlot } from '@#metaName/runtime/rsc/client';
+import { renderRsc } from '@#metaName/runtime/rsc/server';
 export { handleAction } from '@#metaName/runtime/rsc/server';
 
 const handleRequest = async (request, ServerRoot, options) => {
@@ -36,7 +37,7 @@ const handleRequest = async (request, ServerRoot, options) => {
     </ServerRoot>,
     {
       ...options,
-      rscRoot: <options.RSCRoot />,
+      rscRoot: options.rscRoot,
     },
   );
 
@@ -48,7 +49,22 @@ const handleRequest = async (request, ServerRoot, options) => {
 };
 
 export const requestHandler = createRequestHandler(handleRequest, {
-  enableRsc: true,
+  enableRsc: true
+});
+
+const handleRSCRequest = async (request, ServerRoot, options) => {
+  const { serverPayload } = options;
+  const stream = renderRsc({
+    element: options.rscRoot,
+    clientManifest: options.rscClientManifest!,
+  });
+
+  return new Response(stream);
+}
+
+
+export const rscPayloadHandler = createRequestHandler(handleRSCRequest, {
+  enableRsc: true
 });
 `;
 
@@ -67,26 +83,61 @@ export const serverIndex = (options: ServerIndexOptinos) => {
 
 export const entryForCSRWithRSC = ({
   metaName,
+  entryName,
 }: {
   metaName: string;
+  entryName: string;
 }) => {
   return `
-  import App from './AppProxy';
-  import { renderRsc } from '@${metaName}/runtime/rsc/server'
+  import '@${metaName}/runtime/registry/${entryName}';
+  import {
+    createRequestHandler,
+  } from '@${metaName}/runtime/ssr/server';
+  import { renderRsc, processRSCStream } from '@${metaName}/runtime/rsc/server'
   export { handleAction } from '@${metaName}/runtime/rsc/server';
 
+  const handleCSRRender = async (request, ServerRoot, options) => {
+    const rscPayloadStream = renderRsc({
+      element: options.rscRoot,
+      clientManifest: options.rscClientManifest!,
+    });
+    const stream = new ReadableStream({
+      start(controller) {
+        const encoder = new TextEncoder();
 
-  export const rscRequestHandler = ({
-    clientManifest
-  }) => {
-    const stream = renderRsc({
-      element: <App/>,
-      clientManifest,
-    })
+        controller.enqueue(encoder.encode(options.html));
 
-    const response = new Response(stream);
-    return response
+      processRSCStream(rscPayloadStream, controller, encoder)
+        .catch(err => {
+          controller.error(err);
+        });
+      }
+    });
+
+    return new Response(stream, {
+      status: 200,
+      headers: new Headers({
+        'content-type': 'text/html; charset=UTF-8',
+      }),
+    });
   }
+
+  export const renderRscStreamHandler = createRequestHandler(handleCSRRender, {
+    enableRsc: true
+  });
+
+  const handleRequest = async (request, ServerRoot, options) => {
+    const stream = renderRsc({
+            element: options.rscRoot,
+      clientManifest: options.rscClientManifest!,
+    });
+
+    return new Response(stream);
+  }
+
+  export const rscPayloadHandler = createRequestHandler(handleRequest, {
+    enableRsc: true
+  });
 `;
 };