+++ packages/runtime/plugin-runtime/src/router/cli/index.ts
@@ -54,7 +54,7 @@ export const routerPlugin = (): CliPluginFuture<AppTools<'shared'>> => ({
       if ((nestedRoutesEntry || pageRoutesEntry) && !isRouterV5) {
         plugins.push({
           name: 'router',
-          path: `@${metaName}/runtime/router`,
+          path: `@${metaName}/runtime/router/internal`,
           config:
             typeof routerConfig === 'boolean'
               ? { serverBase }
@@ -74,9 +74,10 @@ export const routerPlugin = (): CliPluginFuture<AppTools<'shared'>> => ({
             // react-router v6 is no longer support ie 11
             // so we need to compile these packages to ensure the compatibility
             // https://github.com/remix-run/react-router/commit/f6df0697e1b2064a2b3a12e8b39577326fdd945b
-            /node_modules\/react-router/,
-            /node_modules\/react-router-dom/,
-            /node_modules\/@remix-run\/router/,
+            /[\\/]node_modules[\\/]react-router[\\/]/,
+            /[\\/]node_modules[\\/]react-router-dom[\\/]/,
+            /[\\/]node_modules[\\/]@remix-run[\\/]router[\\/]/,
+            path.resolve(__dirname, '../runtime').replace('cjs', 'esm'),
           ],
           globalVars: {
             'process.env._MODERN_ROUTER_VERSION': 'v6',
@@ -103,7 +104,7 @@ export const routerPlugin = (): CliPluginFuture<AppTools<'shared'>> => ({
       );
       if (!isRouterV5) {
         pluginsExportsUtils.addExport(
-          `export { default as router } from '@${metaName}/runtime/router'`,
+          `export { default as router } from '@${metaName}/runtime/router/internal'`,
         );
       }
     });