+++ packages/runtime/plugin-runtime/src/router/index.ts
@@ -1,3 +1,1 @@
-export { default as router } from './runtime';
-export { default } from './runtime';
 export * from './runtime';