+++ packages/runtime/plugin-runtime/src/router/runtime/index.ts
@@ -1,19 +1,6 @@
 import { useRouteLoaderData as useRouteData } from '@modern-js/runtime-utils/router';
-import { routerPlugin } from './plugin';
-import type { RouterConfig, SingleRouteConfig } from './types';
 
 export * from '@modern-js/runtime-utils/router';
-
-export type { SingleRouteConfig, RouterConfig };
-export { renderRoutes } from './utils';
-
-export { routerPlugin };
-export default routerPlugin;
-
-export { modifyRoutes } from './plugin';
-export type { RouterExtendsHooks } from './hooks';
-
-export * from './withRouter';
 
 export { Link, NavLink } from './PrefetchLink';
 export type { LinkProps, NavLinkProps } from './PrefetchLink';
@@ -28,5 +15,7 @@ export {
   handleRouteModule,
   handleRouteModuleError,
 } from './routeModule';
+
+export * from './withRouter';
 
 export type { LoaderFunction, LoaderFunctionArgs } from './types';