+++ packages/runtime/plugin-runtime/src/router/runtime/internal.ts
@@ -1,0 +1,8 @@
+import { routerPlugin } from './plugin';
+import type { RouterConfig, SingleRouteConfig } from './types';
+export { routerPlugin };
+export default routerPlugin;
+export type { SingleRouteConfig, RouterConfig };
+export type { RouterExtendsHooks } from './hooks';
+export { renderRoutes } from './utils';
+export { modifyRoutes } from './plugin';