+++ packages/runtime/plugin-runtime/src/rsc/client.ts
@@ -1,1 +1,4 @@
 export * from '@modern-js/render/client';
+export const isRedirectResponse = (res: Response) => {
+  return res.headers.get('X-Modernjs-Redirect') != null;
+};