+++ packages/runtime/plugin-runtime/tests/core/createApp.test.tsx
@@ -1,3 +1,6 @@
+/**
+ * @jest-environment jsdom
+ */
 import { render } from '@testing-library/react';
 import React from 'react';
 import { type Plugin, createApp, useRuntimeContext } from '../../src/core';