+++ packages/solutions/app-tools/src/builder/shared/builderPlugins/adapterSSR.ts
@@ -4,7 +4,7 @@ import {
   SERVICE_WORKER_ENVIRONMENT_NAME,
   isHtmlDisabled,
 } from '@modern-js/uni-builder';
-import { fs, isUseSSRBundle } from '@modern-js/utils';
+import { fs, isUseRsc, isUseSSRBundle } from '@modern-js/utils';
 import {
   type RsbuildPlugin,
   type RspackChain,
@@ -35,7 +35,10 @@ export const builderPluginAdapterSSR = <B extends Bundler>(
         server: {
           // the http-compression can't handler stream http.
           // so we disable compress when user use stream ssr temporarily.
-          compress: isStreamingSSR(normalizedConfig) ? false : undefined,
+          compress:
+            isStreamingSSR(normalizedConfig) || isUseRsc(normalizedConfig)
+              ? false
+              : undefined,
         },
       });
     });
@@ -72,7 +75,7 @@ export const builderPluginAdapterSSR = <B extends Bundler>(
           });
         }
 
-        if (isUseSSRBundle(normalizedConfig)) {
+        if (isUseSSRBundle(normalizedConfig) || isUseRsc(normalizedConfig)) {
           await applySSRLoaderEntry(chain, options, isServer);
           applySSRDataLoader(chain, options);
         }
@@ -121,7 +124,7 @@ function applyAsyncChunkHtmlPlugin({
   modernConfig: AppNormalizedConfig<'shared'>;
   HtmlBundlerPlugin: any;
 }) {
-  if (isStreamingSSR(modernConfig)) {
+  if (isStreamingSSR(modernConfig) || isUseRsc(modernConfig)) {
     chain
       .plugin('html-async-chunk')
       .use(HtmlAsyncChunkPlugin, [HtmlBundlerPlugin]);