+++ packages/generator/generators/dependence-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/dependence-generator
 
+## 3.7.37
+
 ## 3.7.36
 
 ## 3.7.35