+++ packages/runtime/plugin-garfish/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @modern-js/plugin-garfish
 
+## 2.68.0
+
+### Patch Changes
+
+- 4cc0a22: feat: support resolve configuration in rsbuild.
+  feat: 支持 rsbuild 的 resolve 配置。
+- Updated dependencies [292d91c]
+- Updated dependencies [4cc0a22]
+  - @modern-js/runtime@2.68.0
+  - @modern-js/plugin-v2@2.68.0
+  - @modern-js/runtime-utils@2.68.0
+  - @modern-js/utils@2.68.0
+
 ## 2.67.11
 
 ### Patch Changes