+++ packages/runtime/plugin-testing/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @modern-js/plugin-testing
 
+## 2.68.0
+
+### Patch Changes
+
+- 4cc0a22: feat: support resolve configuration in rsbuild.
+  feat: 支持 rsbuild 的 resolve 配置。
+- Updated dependencies [292d91c]
+- Updated dependencies [992194b]
+- Updated dependencies [4cc0a22]
+  - @modern-js/runtime@2.68.0
+  - @modern-js/prod-server@2.68.0
+  - @modern-js/bff-core@2.68.0
+  - @modern-js/server-core@2.68.0
+  - @modern-js/babel-compiler@2.68.0
+  - @modern-js/plugin@2.68.0
+  - @modern-js/types@2.68.0
+  - @modern-js/utils@2.68.0
+  - @modern-js/babel-preset@2.68.0
+
 ## 2.67.11
 
 ### Patch Changes