+++ packages/server/utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/server-utils
 
+## 2.68.0
+
+### Patch Changes
+
+- @modern-js/babel-plugin-module-resolver@2.68.0
+- @modern-js/babel-compiler@2.68.0
+- @modern-js/utils@2.68.0
+- @modern-js/babel-preset@2.68.0
+
 ## 2.67.11
 
 ### Patch Changes