+++ packages/toolkit/compiler/babel/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/babel-compiler
 
+## 2.68.0
+
+### Patch Changes
+
+- @modern-js/utils@2.68.0
+
 ## 2.67.11
 
 ### Patch Changes